Putting twilio utils into separate file. Adding alerting for interceptor. Cleaning up interceptor start script.
