+++ new_domain_alert.py
@@ -1,6 +1,6 @@
 import os
 import time
-from twilio.rest import Client
+import twilio_utils
 from pi_hole_admin import PiHoleAdmin
 from unique_domains_windower import UniqueDomainsWindower
 
@@ -18,40 +18,6 @@ def initialize_default_windower(url: str, file_name: str, types: list, only_doma
         windower = UniqueDomainsWindower.deserialize(client, file_name)
 
     return windower
-
-def notify_of_new_domains(domains: list, recipient_phone: str, twilio_phone: str, msg_chars_limit: int=1000, blocked: bool=True):
-    """
-    Send message if a set of new domains has been seen.
-    """
-    if not domains:
-        return
-
-    account_sid = os.environ['TWILIO_ACCOUNT_SID']
-    auth_token = os.environ['TWILIO_AUTH_TOKEN']
-    client = Client(account_sid, auth_token)
-
-    body = f"{'Blocked' if blocked else 'Permitted'} {len(domains)} domains that have not been seen recently:\n"
-
-    i = 0
-
-    while i < len(domains) and len(body) < msg_chars_limit:
-        if len(body) + len(domains[i]) < msg_chars_limit:
-            body += f"{domains[i]}\n"
-
-            i += 1
-        else:
-            break
-
-    if i != len(domains):
-        body += "..."
-
-    message = client.messages.create(
-        body=body,
-        from_=twilio_phone,
-        to=recipient_phone
-    )
-
-    print(f"Sending message with SID: {message.sid}")
 
 def main():
     start = time.time()
@@ -78,7 +44,7 @@ def main():
 
     print(f"Found blocked domains {previously_unseen_blocked_domains}")
 
-    notify_of_new_domains(previously_unseen_blocked_domains, os.environ["ADMIN_PHONE"], os.environ["TWILIO_PHONE"], blocked=True)
+    twilio_utils.notify_of_new_domains(previously_unseen_blocked_domains, os.environ["ADMIN_PHONE"], os.environ["TWILIO_PHONE"], blocked=True)
     
     print(f"Finished blacklist assessment in {time.time() - start} sec")
 
@@ -90,7 +56,7 @@ def main():
 
     print(f"Found permitted domains {previously_unseen_permitted_domains}")
 
-    notify_of_new_domains(previously_unseen_permitted_domains, os.environ["ADMIN_PHONE"], os.environ["TWILIO_PHONE"], blocked=False)
+    twilio_utils.notify_of_new_domains(previously_unseen_permitted_domains, os.environ["ADMIN_PHONE"], os.environ["TWILIO_PHONE"], blocked=False)
 
     print(f"Finished whitelist assessment in {time.time() - start} sec")
